feat: expose base router without default mapper
Use `import Router from 'cerebral-module-router/base'` to use router with custom mapper.
You can pass object with `map` and `stringify` properties on it to `options.mapper`. See `url-mapper` package for details.

BREAKING CHANGE:
Remove deprecated way to enable queries in default router through `options.mapper`.
Use `options.query` directly instead.
